+++ client/src/pages/studyCollectionsDetails/StudyCollectionDetails.js
@@ -23,16 +23,20 @@ import {
   Dropdown,
   Row
 } from "react-bootstrap";
-import React from "react";
+import React, {useContext} from "react";
 import {Edit2, File, Menu} from "react-feather";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {faEdit} from "@fortawesome/free-solid-svg-icons";
+import {faEdit, faTrash} from "@fortawesome/free-solid-svg-icons";
 import ToolkitProvider, {Search} from "react-bootstrap-table2-toolkit";
 import BootstrapTable from "react-bootstrap-table-next";
 import paginationFactory from "react-bootstrap-table2-paginator";
 import {StatusBadge} from "../../common/status";
 import PropTypes from "prop-types";
 import StudyUpdateModal from "./StudyUpdateModal";
+import swal from "sweetalert";
+import axios from "axios";
+import {useNavigate} from "react-router-dom";
+import NotyfContext from "../../context/NotyfContext";
 
 const StudyCollectionDetailsHeader = ({collection}) => {
   return (
@@ -66,6 +70,8 @@ const StudyCollectionDetails = ({
 }) => {
 
   const [modalIsOpen, setModalIsOpen] = React.useState(false);
+  const navigate = useNavigate();
+  const notyf = useContext(NotyfContext);
 
   const handleStudiesUpdate = (studies) => {
     handleUpdateCollection({
@@ -73,6 +79,29 @@ const StudyCollectionDetails = ({
       studies: studies
     });
   };
+
+  const handleCollectionDelete = () => {
+    swal({
+      title: "Are you sure you want to delete this collection?",
+      icon: "warning",
+      buttons: true
+    })
+    .then(val => {
+      if (val) {
+        axios.delete("/api/internal/studycollection/" + collection.id)
+        .then(response => {
+          navigate("/collections");
+        })
+        .catch(error => {
+          console.error(error);
+          notyf.open({
+            type: "error",
+            message: "Error deleting collection."
+          })
+        })
+      }
+    });
+  }
 
   const columns = [
     {
@@ -182,24 +211,6 @@ const StudyCollectionDetails = ({
 
           ) : ''
     },
-    // {
-    //   dataField: "remove",
-    //   text: "Remove",
-    //   sort: false,
-    //   searchable: false,
-    //   headerStyle: {width: '10%'},
-    //   formatter: (c, d, i, x) => {
-    //     return (
-    //         <div>
-    //           <a className="text-danger"
-    //              title={"Remove study from collection"}
-    //              onClick={() => handleRemoveStudy(d.id)}>
-    //             <XCircle className="align-middle me-1" size={18}/>
-    //           </a>
-    //         </div>
-    //     )
-    //   }
-    // },
     {
       dataField: 'search',
       text: 'Search',
@@ -269,9 +280,13 @@ const StudyCollectionDetails = ({
 
                       <Dropdown.Item
                           href={"/collection/" + collection.id + "/edit"}>
-                        <FontAwesomeIcon icon={faEdit}/>
-                        &nbsp;
+                        <FontAwesomeIcon className={"me-2"} icon={faEdit}/>
                         Edit
+                      </Dropdown.Item>
+
+                      <Dropdown.Item onClick={handleCollectionDelete}>
+                        <FontAwesomeIcon className={"me-2"} icon={faTrash}/>
+                        Delete
                       </Dropdown.Item>
 
                     </Dropdown.Menu>